requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
Pillow
pytz
sqlparse
python-dotenv
PyMySQL
requests
stripe
